Backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
yfinance==0.2.36
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
python-dotenv==1.0.1
requests==2.31.0
requests-cache==1.2.0
websockets==12.0
langchain
langchain-core
langchain-groq
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0